- Removed SDL_Renderer from source tree and officially deprecated it. Refs #61. - Cleaned unnecessary c'tor from Filesystem. - Removed WM code from Renderer. - Pulled all texture management code from Renderer and moved to Image instead. Refs #66. - Minor changes to ShaderManager. - Image is now responsible for generating and managing texture ID's. Refs #61 and #66. Fixes #33. - Updated logger printing. Cosmetic changes only.

+++ core/src/Resources/Sprite.cpp
@@ -155,7 +155,7 @@ void Sprite::update(int x, int y)
 
 	if(!mPaused && (frame.frameDelay() != FRAME_PAUSE))
 	{
-		while(frame.frameDelay() > 0 && mTimer.accumulator() >= frame.frameDelay())
+		while(frame.frameDelay() > 0 && static_cast<int>(mTimer.accumulator()) >= frame.frameDelay())
 		{
 			mTimer.adjust_accumulator(frame.frameDelay());
 			mCurrentFrame++;
